use BrowserDialog objects in MainWindow
BrowserDialog objects are parentless to allow the user to reposition
away from the fractal and see applied changes.

Current formula and file are selected when the dialog is opened.

A new dialog is created on each use. Converting to an object level
dialog that is hidden between uses would require a review of the dialog
and browser_model.