improve video generation
Speed up by removing Python threading and solely using the signals
emitted by gtkfractal.

Clarify that a video is made up of keyframes and transitions
consisting of a number of frames defined by the duration parameter.
By default the keyframe is shown for one frame but this is
configurable using the stopped parameter.