remove fract4d_stdlib
Complicated story: at runtime, we load a generated shared
object containing the code for the fractal. We want that to
be able to call some helper functions that are contained in
the fract4dc module. That module is also dynamically loaded,
and by default Python loads it without specifying RTLD_GLOBAL.
That means functions in it are not visible to other dynamically
loaded modules, which we need.

Originally I resolved this with a terrible hack: create another .so
containing the code I wanted, and dlopen() that specifically
with RTLD_GLOBAL. Looks like there is an easier fix, using
sys.setdlopenflags, which I never realized.

This simplifies things by removing that extra complication. Sweet!

TODO: need to adjust the c++ example code to match.